Importing Selenium Goes Wrong
I was debugging and solidifying the script automation. I also did output_debugged the counter and turned it to a for loop instead of a static output print.

Moved on to importing the selenium module. 80% of my time was contributed finding out the problem of importing the selenium. After running out of time. I figured out i was importing the Selenium the wrong way. Next time, I have to Import it properly.