boot site error fixed
the script was not able to boot the site properly and input the the email with 'sendKeys'. I failed to realize i was calling the function not even being declared first, hence the boot site fails.

the script fails to 'sendKeys' the input password not because of the improper element input used but rather it 'sendKeys' input for the password while the site is loading. I added a delayed feature before executing the line for input password and not it works.

Boot site launched successfully and Sending Credentials is now Doable.