School Site Automated Logged in Successfully
The problem was i copied the link from a incognito window in the script. That is why security measure triggered thus failing me to log in despite the automated log in.

I deleted the link. copied the link from a normal chrome window and pasted in my script and added an --incognito argument. and it now works. as for now i am doing the same with Notion site and my Github.

more to be followed...